Typechecking half works.  Exciting.
I'm going to add an "illumination" pass that unshadows all names via
renaming.  Typechecking will proceed after this pass, and so can rely
on the uniqueness of identifiers.  I will have to attach debug
information later in order to "eclipse" the names back to what they
were in the source code in case of error.

Added a .ghci file and some depedencies to pretty-print the damn
syntax trees automatically.  These will be removed at a later date
when I'm not doing tests by manually looking at the AST that comes
out.